Clean up unused behaviors and documentation
Without this patch there are a bunch of methods left over from the
previous approach of trying to modify the application wrapper script.
This is a problem because these methods are not obviously different and
distinct behaviors from the approach of modifying the default system
confdir and vardir inside of Puppet itself.

This patch addresses the problem by removing all of the previous code
and updating the documentation appropriately.